% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{load_sounds}
\alias{load_sounds}
\title{Load sound files}
\usage{
load_sounds(sound_files = getOption("celebrate.sounds"))
}
\arguments{
\item{sound_files}{list of names of WAV files}
}
\description{
Load sound files
}
